requirements.txt
setuptools==75.1.0
flask-mongoengine==1.0.0
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
gunicorn==23.0.0
openpyxl==3.1.5
PyJWT==2.7.0
pymongo==4.10.1
flask_jwt_extended==4.7.1
marshmallow==3.23.2
razorpay==1.4.2
cryptography==44.0.0
python-dotenv==1.0.1
python-dateutil==2.9.0